LayoutTests/http/tests/websocket/tests/hybi/simple-wss.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../../js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Simple secure Web Socket test");

window.jsTestIsAsync = true;

if (window.testRunner)
    testRunner.setAllowsAnySSLCertificate(true);

function endTest()
{
    clearTimeout(timeoutID);
    finishJSTest();
}

var ws = new WebSocket("wss://127.0.0.1:9323/websocket/tests/hybi/simple");
debug("Created a socket to '" + ws.url + "'; readyState " + ws.readyState + ".");

ws.onopen = function()
{
    debug("Connected; readyState " + ws.readyState);
};

ws.onmessage = function(messageEvent)
{
    debug("Received: '" + messageEvent.data + "'; readyState " + ws.readyState);
};

ws.onclose = function()
{
    debug("Closed; readyState " + ws.readyState + ".");
    endTest();
};

function timeOutCallback()
{
    debug("Timed out in state: " + ws.readyState);
    endTest();
}

var timeoutID = setTimeout(timeOutCallback, 3000);

</script>
<script src="../../../../js-test-resources/js-test-post.js"></script>
</body>
</html>
