LayoutTests/platform/glib/http/tests/websocket/tests/hybi/simple-wss-expected.txt

Simple secure Web Socket test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Created a socket to 'wss://127.0.0.1:9323/websocket/tests/hybi/simple'; readyState 0.
Connected; readyState 1
Received: 'Hello from Simple WSH.'; readyState 1
Timed out in state: 1
PASS successfullyParsed is true

TEST COMPLETE
